move transaction details
