+++ src/test/java/xyz/npgw/test/page/common/table/TransactionsTableComponent.java
@@ -5,7 +5,7 @@ import com.microsoft.playwright.Page;
 import com.microsoft.playwright.options.AriaRole;
 import io.qameta.allure.Step;
 import xyz.npgw.test.page.TransactionsPage;
-import xyz.npgw.test.page.dialog.TransactionDetailsDialog;
+import xyz.npgw.test.page.dialog.transactions.TransactionDetailsDialog;
 import xyz.npgw.test.page.dialog.transactions.RefundTransactionDialog;
 
 import java.net.URLDecoder;
